Continued box2d porting. Objects not bouncing for some reason.

+++ spacegame/src/uk/digitalsquid/spacegame/spaceitem/items/Planet.java
@@ -19,13 +19,13 @@ import uk.digitalsquid.spacegamelib.spaceitem.Gravitable;
 public class Planet extends Gravitable
 {
 	protected static final PlanetType[] PLANET_TYPES = {
-		new PlanetType(Type.nobounce1,	FgType.image,	R.drawable.planet1,		0.2, 0.5, .30f, 2.50f, null),
-		new PlanetType(Type.sticky1,	FgType.image,	R.drawable.planet5,		0,   0.9, .40f, 2.50f, null),
-		new PlanetType(Type.bounce2,	FgType.image,	R.drawable.planet2,		1,   0.1, .20f, 2.00f, null),
-		new PlanetType(Type.n1,			FgType.image,	R.drawable.planet3,		0.6, 0.6, .20f, 2.00f, null),
-		new PlanetType(Type.n2,			FgType.image,	R.drawable.planet4,		0.7, 0.7, .20f, 2.00f, null),
-		new PlanetType(Type.n3,			FgType.image,	R.drawable.planet6,		0.8, 0.8, .20f, 2.00f, null),
-		new PlanetType(Type.bounce1,	FgType.image,	R.drawable.planetbouncy,1.3, 0.8, .20f, 1.50f, null)
+		new PlanetType(Type.nobounce1,	FgType.image,	R.drawable.planet1,		0.2f, 0.5f, .30f, 2.50f, null),
+		new PlanetType(Type.sticky1,	FgType.image,	R.drawable.planet5,		0f,   0.9f, .40f, 2.50f, null),
+		new PlanetType(Type.bounce2,	FgType.image,	R.drawable.planet2,		1f,   0.1f, .20f, 2.00f, null),
+		new PlanetType(Type.n1,			FgType.image,	R.drawable.planet3,		0.6f, 0.6f, .20f, 2.00f, null),
+		new PlanetType(Type.n2,			FgType.image,	R.drawable.planet4,		0.7f, 0.7f, .20f, 2.00f, null),
+		new PlanetType(Type.n3,			FgType.image,	R.drawable.planet6,		0.8f, 0.8f, .20f, 2.00f, null),
+		new PlanetType(Type.bounce1,	FgType.image,	R.drawable.planetbouncy,1.3f, 0.8f, .20f, 1.50f, null)
 	};
 	
 	protected PlanetType type = PLANET_TYPES[0];
@@ -40,6 +40,7 @@ public class Planet extends Gravitable
 		super(context, coord, 1, getDensityForId(typeId), radius, BodyType.STATIC);
 		
 		fixture.setRestitution(getBouncinessForId(typeId));
+		// fixture.setFriction(CompuFuncs.TrimMinMax(1 - getBouncinessForId(typeId), 0, 1));
 		
 		for(PlanetType t : PLANET_TYPES) {
 			if(t.typeId == typeId) {
@@ -120,11 +121,11 @@ public class Planet extends Gravitable
 		
 		public final float minSize, maxSize;
 		
-		public final double bounciness, density;
+		public final float bounciness, density;
 		
 		public final PaintDesc colour1;
 		
-		protected PlanetType(int typeId, FgType fg, int fileId2, double bounciness, double density, float minSize, float maxSize, PaintDesc colour1) {
+		protected PlanetType(int typeId, FgType fg, int fileId2, float bounciness, float density, float minSize, float maxSize, PaintDesc colour1) {
 			this.typeId = typeId;
 			this.fileId2 = fileId2;
 			this.minSize = minSize;
